Add files via upload
I can still see the password of the deleted code, so I changed the password.

The original use of the code was to gather posts to train a Tensor Flow model to try and make a 'Republican' Twitter bot that automatically posted random Tweets. I had some success but it mostly Tweeted nonsense. 

I wanted to make this to see how easy it could be to make bots because of the upcoming election and the problem of mis/disinformation.